Add passwd/group copy/clone functions.
I didn't know if this kind of thing would be necessary, but
nfsutils does seem to do that thing where it allocates
a passwd struct and then returns it with the expectation
that the caller or some later function will call free() on
that passwd struct. Trying to refactor that code to take
advantage of opportunistic stack allocation might be asking
a bit too much. It's probably better to just malloc this
stuff and keep things as untouched as possible.