More robust memory management; stronger guarantees.
A lot of this is just comment/documentation improvements to
explain what was already there, but there were also a few code
changes that make the stronger guarantees possible.

The intent is to make the caller's side less messy by making
it possible to do things like call the *_cleanup functions
unconditionally, or call the *_result functions after failed
queries and get predictable (NULL/blank) outputs. The caller
should not endure the burden of frivolous error handling
and unnecessary pipeline interventions.

There was also a bugfix in there regarding the realloc loop
never seeming to return a value when succeeding. I'm not
sure why that ever worked, but it was probably undefined
behavior that just happened to do the right thing on my
system. No matter, it's fixed now, and I intend that it
ALWAYS does the right thing.