Added convenience functions and error printer.
This makes things very simple for cases where only uid+gid or gid
are needed, and the strings would never be used.

It also offers a "just give me a struct" version that expects the
caller to call `free` on the output struct.

Both variants return ENOENT to indicate non-erroring failure to
find an entry (e.g. the entry doesn't exist in the passwd/group file).
This makes it easier to write
  if(err) print_errors(...);
  else do_something(pwgrp_thing);
style code.

There is also an error printer that prints all possible errors
generated by this family of functions. Even ENOMEM. Without malloc.
Maybe. (It depends on what libc|syslog do with plain strings.)